+++ etc/ci_test.sh
@@ -14,8 +14,13 @@ CFLAGS="--coverage" LDFLAGS="--coverage" FORCE_JULIAINTERFACE_COMPILATION=true $
 ${GAP} makedoc.g
 ${GAP} --cover ../../coverage/JuliaInterface.coverage tst/testall.g || AnyFailures=Yes
 gcov -o gen/src/ src/*.c*
-# Remove JuliaInterface with coverage instrumentation
-make clean
+# We should delete the coverage instrumentation in JuliaInterface now
+# by calling `make clean`. However, this does not work because the
+# `gaproot_for_building` only exists while GAP.jl runs.
+# So we instead start a new GAP.jl session that forces a recompilation
+# of JuliaInterface with default settings, thus overwriting the
+# coverage instrumentation.
+FORCE_JULIAINTERFACE_COMPILATION=true ${GAP} --nointeract
 cd ../..
 
 #