+++ src/packages.jl
@@ -17,6 +17,7 @@ import Pidfile
 import Scratch: @get_scratch!
 import ...GAP: disable_error_handler, Globals, GapObj, replace_global!, RNamObj, Wrappers
 import ...GAP: gap_pkg_jlls, Compat, GAP_VERSION, GAP_jll, GAP_lib_jll
+import ...GAP.Setup: gaproot_for_building
 
 gap_packages_rootdir() = @get_scratch!("gap_packagedir_v$(GAP_VERSION.major).$(GAP_VERSION.minor)")
 
@@ -254,6 +255,9 @@ function install(spec::String, version::String = "";
     # point PackageManager to the given pkg dir
     Globals.PKGMAN_CustomPackageDir = GapObj(pkgdir)
     mkpath(pkgdir)
+    # inject our custom sysinfo.gap into the package manager
+    Globals.PKGMAN_Sysinfo = GapObj(joinpath(gaproot_for_building(), "sysinfo.gap"))
+
     # pidlock timeout: 300 seconds = 5 minutes should suffice; about the
     # slowest packages to install are Semigroups and NormalizInterface.
     Pidfile.mkpidlock("$pkgdir.lock"; stale_age=300) do
@@ -397,6 +401,10 @@ function build(name::String; quiet::Bool = false,
 
   gname = GapObj(name)
   Globals.TestPackageAvailability(gname) != Globals.fail && return true # already available, build not necessary
+
+  # inject our custom sysinfo.gap into the package manager
+  Globals.PKGMAN_Sysinfo = GapObj(joinpath(gaproot_for_building(), "sysinfo.gap"))
+
   allinfo = collect(Globals.PackageInfo(gname))
   userinfo = filter(info -> startswith(String(info.InstallationPath), pkgdir), allinfo)
   isempty(allinfo) && error("package not found")