+++ test/packages.jl
@@ -22,6 +22,11 @@
     @test ! isempty(GAP.Packages.locate_package("FGA"))
     @test GAP.Packages.remove("fga", interactive = false)
 
+    # Test the installation of a package with a kernel extension, but without a jll
+    @test GAP.Packages.install("https://github.com/gap-packages/RegisterPackageTNUMDemo/releases/download/v0.4/RegisterPackageTNUMDemo-0.4.tar.gz", interactive = false)
+    @test GAP.Packages.load("RegisterPackageTNUMDemo")
+    @test GAP.Packages.remove("RegisterPackageTNUMDemo", interactive = false)
+
 #    pkgdir = mktempdir()
 #    @test GAP.Packages.install("fga", interactive = false, pkgdir = pkgdir)
 #    @test GAP.Packages.remove("fga", interactive = false, pkgdir = pkgdir)